Render qmd as pptx

Render a qmd as pptx, applying the Atorus theme.

Usage

render_pptx(path, date = Sys.Date(), quiet = TRUE, ...)

Arguments

path

The path to the qmd file to render. By default, a new file will be created in the same directory as path, with extension .pptx.

date

A date to include on the title slide of the deck. Set this to NULL if you do not wish to include a date.

quiet

Whether to suppress output during the rendering process. Passed on to quarto::quarto_render().

...

Arguments passed on to quarto::quarto_render

output_file

The name of the output file. If using NULL, the output filename will be based on the filename for the input file. output_file is mapped to the --output option flag of the quarto CLI. It is expected to be a filename only, not a path, relative or absolute.

execute

Whether to execute embedded code chunks.

execute_params

A list of named parameters that override custom params specified within the YAML front-matter.

execute_daemon

Keep Jupyter kernel alive (defaults to 300 seconds). Note this option is only applicable for rendering Jupyter notebooks or Jupyter markdown.

execute_daemon_restart

Restart keepalive Jupyter kernel before render. Note this option is only applicable for rendering Jupyter notebooks or Jupyter markdown.

execute_debug

Show debug output for Jupyter kernel.

use_freezer

Force use of frozen computations for an incremental file render.

cache_refresh

Force refresh of execution cache.

metadata

An optional named list used to override YAML metadata. It will be passed as a YAML file to --metadata-file CLI flag. This will be merged over metadata-file options if both are specified.

metadata_file

A yaml file passed to --metadata-file CLI flags to override metadata. This will be merged with metadata if both are specified, with low precedence on metadata options.

debug

Leave intermediate files in place after render.

profile

Quarto project profile(s) to use. Either a character vector of profile names or NULL to use the default profile.

quarto_args

Character vector of other quarto CLI arguments to append to the Quarto command executed by this function. This is mainly intended for advanced usage and useful for CLI arguments which are not yet mirrored in a dedicated parameter of this R function. See quarto render --help for options.

as_job

Render as an RStudio background job. Default is "auto", which will render individual documents normally and projects as background jobs. Use the quarto.render_as_job R option to control the default globally.

Value

NULL invisibly (this function is called to generate the pptx as a side effect).